# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: FieldForce Live — Admin Panel API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /onboarding/{token}:
    get:
      operationId: getOnboardingByToken
      tags: [onboarding]
      summary: Resolve an onboarding token to its landing-page data
      parameters:
        - name: token
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Onboarding info
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OnboardingLanding"
        "404":
          description: Invalid token
  /onboarding/{token}/consent:
    post:
      operationId: submitOnboardingConsent
      tags: [onboarding]
      summary: Record consent and activate the user
      parameters:
        - name: token
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Consent recorded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OnboardingLanding"

  # ---------- Ingestion (mobile-facing) ----------
...
  /user/status:
    post:
      operationId: postUserStatus
      tags: [mobile-status]
      summary: Update field user live status (IDLE or BUSY)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserStatusInput"
      responses:
        "200":
          description: Status updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserStatusResult"
  /user/emergency:
    post:
      operationId: postUserEmergency
      tags: [mobile-status]
      summary: Raise or clear a user emergency
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserEmergencyInput"
      responses:
        "200":
          description: Emergency state updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserEmergencyResult"
  /user/visit/{visitStopId}/disposition:
    post:
      operationId: closeVisitStop
      tags: [mobile-status]
      summary: Close a visit stop with a disposition
      parameters:
        - name: visitStopId
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CloseVisitStopInput"
      responses:
        "200":
          description: Visit stop closed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VisitStop"
        "404":
          description: Not found
  /user/push-token:
    post:
      operationId: registerPushToken
      tags: [mobile-status]
      summary: Register a push notification token (stub — logs only, no storage)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PushTokenInput"
      responses:
        "204":
          description: Registered

  # ---------- Simulator ----------
...
components:
  schemas:
...
    UserStatus:
      type: string
      enum: [INVITED, ACTIVE, SUSPENDED]
...
    VisitStopPriority:
      type: string
      enum: [P1, P2, P3]
    VisitStopStatus:
      type: string
      enum: [PENDING, EN_ROUTE, REACHED, COMPLETED, SKIPPED]
    VisitStopInputType:
      type: string
      enum: [ADDRESS, PIN, LATLNG]
...
    PushPlatform:
      type: string
      enum: [FCM, APNS, HMS]
...
    OnboardingLanding:
      type: object
      required: [userFirstName, status, consentGivenAt, iosStoreUrl, androidStoreUrl, huaweiStoreUrl]
      properties:
        userFirstName: { type: string }
        status:
          $ref: "#/components/schemas/UserStatus"
        consentGivenAt: { type: ["string", "null"], format: date-time }
        iosStoreUrl: { type: string }
        androidStoreUrl: { type: string }
        huaweiStoreUrl: { type: string }
...
    VisitStop:
      type: object
      required:
        [
          id,
          userId,
          dayPlanId,
          visitDate,
          sequence,
          priority,
          customerCode,
          label,
          inputType,
          rawInput,
          latitude,
          longitude,
          status,
        ]
      properties:
        id: { type: integer }
        userId: { type: integer }
        dayPlanId: { type: integer }
        visitDate: { type: string, format: date }
        sequence: { type: ["integer", "null"] }
        priority:
          $ref: "#/components/schemas/VisitStopPriority"
        customerCode: { type: string }
        label: { type: ["string", "null"] }
        inputType:
          $ref: "#/components/schemas/VisitStopInputType"
        rawInput: { type: string }
        latitude: { type: number }
        longitude: { type: number }
        status:
          $ref: "#/components/schemas/VisitStopStatus"
        plannedArrivalAt: { type: ["string", "null"], format: date-time }
        actualArrivalAt: { type: ["string", "null"], format: date-time }
        contactName: { type: ["string", "null"] }
        contactPhone: { type: ["string", "null"] }
        reachedAt: { type: ["string", "null"], format: date-time }
        startedAt: { type: ["string", "null"], format: date-time }
        closedAt: { type: ["string", "null"], format: date-time }
        dispositionId: { type: ["integer", "null"] }
        notes: { type: ["string", "null"] }
...
    UserStatusInput:
      type: object
      required: [userId, status, lat, lng, at]
      properties:
        userId: { type: integer }
        status: { type: string, enum: [IDLE, BUSY] }
        visitStopId: { type: integer }
        lat: { type: number }
        lng: { type: number }
        at: { type: string, format: date-time }
    UserStatusResult:
      type: object
      required: [userId, status]
      properties:
        userId: { type: integer }
        status: { type: string, enum: [IDLE, BUSY] }
    UserEmergencyInput:
      type: object
      required: [userId, active, lat, lng]
      properties:
        userId: { type: integer }
        active: { type: boolean }
        lat: { type: number }
        lng: { type: number }
    UserEmergencyResult:
      type: object
      required: [userId, emergencyActive]
      properties:
        userId: { type: integer }
        emergencyActive: { type: boolean }
    CloseVisitStopInput:
      type: object
      required: [dispositionId, closedAt]
      properties:
        dispositionId: { type: integer }
        notes: { type: string, maxLength: 250 }
        reachedAt: { type: string, format: date-time }
        startedAt: { type: string, format: date-time }
        closedAt: { type: string, format: date-time }
        lat: { type: number }
        lng: { type: number }
    PushTokenInput:
      type: object
      required: [userId, token, platform]
      properties:
        userId: { type: integer }
        token: { type: string }
        platform:
          $ref: "#/components/schemas/PushPlatform"