# lib/api-zod/src/generated/types/pushPlatform.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * FieldForce Live — Admin Panel API
 * OpenAPI spec version: 0.1.0
 */

export type PushPlatform = typeof PushPlatform[keyof typeof PushPlatform];


export const PushPlatform = {
  FCM: 'FCM',
  APNS: 'APNS',
  HMS: 'HMS',
} as const;