# lib/api-zod/src/generated/types/userEmergencyResult.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * FieldForce Live — Admin Panel API
 * OpenAPI spec version: 0.1.0
 */

export interface UserEmergencyResult {
  userId: number;
  emergencyActive: boolean;
}